solidity refactor in progres
